requirements.txt
openai-whisper==20231117
spacy==3.7.4
torch==2.2.2
numpy==1.26.4
tqdm==4.66.4
regex==2024.11.6
requests==2.32.3
